refactor(documents): use hidden attribute and state machine pattern for connection status
Replace inline style.display manipulation with the semantic hidden HTML
attribute for better accessibility. Consolidate repetitive show/hide
logic into a single activateState method driven by a ConnectionState
type union, and align Stimulus target names (live/offline/recovered)
between controller and template.
