+++ modules/documents/app/components/documents/show_edit_view/page_header/connection_status_component.html.erb
@@ -39,36 +39,32 @@
     }
   ) do
     flex_layout(align_items: :center) do |flex|
-      flex.with_column(mr: 3, data: { normal: "" }) do
+      flex.with_column(mr: 3, data: { "documents--connection-status-target": "live" }) do
         render(Documents::ShowEditView::PageHeader::LiveUsersComponent.new)
       end
 
-      flex.with_column(data: { normal: "" }) do
+      flex.with_column(data: { "documents--connection-status-target": "live" }) do
         render(Documents::ShowEditView::PageHeader::LiveSavedAtComponent.new(model))
       end
 
-      flex.with_column do
-        tag.span(data: { "documents--connection-status-target": "offline" }, style: "display:none") do
-          flex_layout(align_items: :center) do |row|
-            row.with_column(mr: 1) do
-              render(Primer::Beta::Octicon.new(icon: :alert, color: :attention, size: :small))
-            end
-            row.with_column do
-              render(Primer::Beta::Text.new(color: :attention)) { t("documents.info_line.currently_offline") }
-            end
+      flex.with_column(data: { "documents--connection-status-target": "offline" }, hidden: true) do
+        flex_layout(align_items: :center) do |row|
+          row.with_column(mr: 1) do
+            render(Primer::Beta::Octicon.new(icon: :alert, color: :attention, size: :small))
+          end
+          row.with_column do
+            render(Primer::Beta::Text.new(color: :attention)) { t("documents.info_line.currently_offline") }
           end
         end
       end
 
-      flex.with_column do
-        tag.span(data: { "documents--connection-status-target": "recovered" }, style: "display:none") do
-          flex_layout(align_items: :center) do |row|
-            row.with_column(mr: 1) do
-              render(Primer::Beta::Octicon.new(icon: :"check-circle", color: :success, size: :small))
-            end
-            row.with_column do
-              render(Primer::Beta::Text.new(color: :success)) { t("documents.info_line.connection_reestablished") }
-            end
+      flex.with_column(data: { "documents--connection-status-target": "recovered" }, hidden: true) do
+        flex_layout(align_items: :center) do |row|
+          row.with_column(mr: 1) do
+            render(Primer::Beta::Octicon.new(icon: :"check-circle", color: :success, size: :small))
+          end
+          row.with_column do
+            render(Primer::Beta::Text.new(color: :success)) { t("documents.info_line.connection_reestablished") }
           end
         end
       end
